+++ examples/shop-with-blog/client/src/pages/shop/Account/Orders/Order.js
@@ -1,9 +1,10 @@
 import React from 'react';
+import { OrderQuery } from '@deity/falcon-shop-data';
 import { H1, Text, Divider, Box, FlexLayout, GridLayout } from '@deity/falcon-ui';
 import { toGridTemplate, FormattedDate, AddressDetails } from '@deity/falcon-ui-kit';
 import { LocaleProvider } from '@deity/falcon-front-kit';
 import { I18n, T } from '@deity/falcon-i18n';
-import { GetOrderQuery, TotalRow, OrderItemSummary } from '@deity/falcon-ecommerce-uikit';
+import { TotalRow, OrderItemSummary } from '@deity/falcon-ecommerce-uikit';
 
 const orderLayoutArea = {
   items: 'checkout',
@@ -35,7 +36,7 @@ const Order = ({ match }) => {
   const id = parseInt(match.params.id, 10);
 
   return (
-    <GetOrderQuery variables={{ id }}>
+    <OrderQuery variables={{ id }}>
       {({ order }) => (
         <GridLayout mb="md" gridGap="md">
           <H1>
@@ -106,7 +107,7 @@ const Order = ({ match }) => {
           </Box>
         </GridLayout>
       )}
-    </GetOrderQuery>
+    </OrderQuery>
   );
 };
