+++ packages/falcon-shop-data/src/Order/index.ts
@@ -1,0 +1,2 @@
+export * from './OrderQuery';
+export * from './LastOrderQuery';
